requirements.txt
aiohttp==3.10.10
anthropic==0.45.2
colorama==0.4.6
discord.py==2.4.0
fastapi==0.115.8
fuzzywuzzy==0.18.0
numpy==2.2.2
ollama==0.4.7
openai==1.61.1
opencv_python==4.10.0.84
Pillow==11.1.0
pydantic==2.10.6
PyGithub==2.5.0
pytest==8.3.4
python-dotenv==1.0.1
PyYAML==6.0.2
tiktoken==0.8.0
uvicorn==0.34.0
